Fix formatting issues in API calls, enhance file input for PDFs, and improve customer approval logic

+++ src/pages/contact-groups/components/delete-dialog.tsx
@@ -14,14 +14,19 @@ interface DeleteNumberDialogProps {
     id: any
     name: string | any
     onClose: () => void
+    type:"number"|"phonebook"
 }
 
-const DeleteNumberDialog = ({ id, name, onClose }: DeleteNumberDialogProps) => {
+const DeleteNumberDialog = ({ id, name, onClose,type  }: DeleteNumberDialogProps) => {
     const queryClient = useQueryClient()
 
     const mutation = useMutation({
         mutationFn: async () => {
-            return await contactService.deletePhoneBookNumber(id)
+            if (type === "number") {
+                return await contactService.deletePhoneBookNumber(id)
+            }else if(type === "phonebook") {
+                 //soft delete phonebook
+            }
         },
         onSuccess: () => {
             toast.success('Number deleted successfully')
@@ -42,9 +47,9 @@ const DeleteNumberDialog = ({ id, name, onClose }: DeleteNumberDialogProps) => {
         <Dialog open={true} onOpenChange={onClose}>
             <DialogContent>
                 <DialogHeader>
-                    <DialogTitle>Delete Phonebook number</DialogTitle>
+                    <DialogTitle>{type === "number" ?"Delete Phonebook number": "Delete Phonebook"}</DialogTitle>
                     <DialogDescription>
-                        Are you sure you want to delete this number <strong>{name}</strong>? This action cannot be undone.
+                        Are you sure you want to delete this {type ===  "number" ? "number": "phonebook"} <strong>{name}</strong>? This action cannot be undone.
                     </DialogDescription>
                 </DialogHeader>
